requirements.txt

Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Pillow==11.2.1
PyTelegramBotAPI==4.27.0
SQLAlchemy==2.0.28
gunicorn==23.0.0
mutagen==1.47.0
psutil==7.0.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
